+++ Application/ApiModels/ProductApiModel.cs
@@ -5,7 +5,7 @@ public class ProductApiModel
     public ProductApiModel(
         int id,
         string name,
-        float price,
+        decimal price,
         string description,
         DateTime dateCreated,
         List<ImageApiModel> images)
@@ -20,7 +20,7 @@ public class ProductApiModel
 
     public int Id { get; private set; }
     public string Name { get; set; }
-    public float Price { get; set; }
+    public decimal Price { get; set; }
     public string Description { get; set; }
     public DateTime DateCreated { get; private set; }
     public List<ImageApiModel> Images { get; set; }
