+++ Domain/Models/Order.cs
@@ -4,7 +4,7 @@ using Domain.ValueObjects.OrderItem;
 namespace Domain.Models;
 public class Order
 {
-    public Order(int id, float total, DateTime dateCreated, DateTime dateFinished, List<OrderItem> orderItems, OrderStatus status)
+    public Order(int id, decimal total, DateTime dateCreated, DateTime dateFinished, List<OrderItem> orderItems, OrderStatus status)
     {
         Id = id;
         Total = total;
@@ -15,7 +15,7 @@ public class Order
     }
 
     public int Id { get; private set; }
-    public float Total { get; set; }
+    public decimal Total { get; set; }
     public OrderStatus Status { get; set; }
     public DateTime DateCreated { get; private set; }
     public DateTime DateFinished { get; set; }
