+++ Domain/Models/ProductHistory.cs
@@ -6,7 +6,7 @@ public class ProductHistory
         int id,
         string name,
         List<string> images,
-        float price,
+        decimal price,
         int productId,
         DateTime validFrom,
         DateTime validTo,
@@ -26,7 +26,7 @@ public class ProductHistory
     public string Name { get; set; }
     public List<string> Images { get; set; }
     public string Description { get; set; }
-    public float Price { get; set; }
+    public decimal Price { get; set; }
     public int ProductId { get; set; }
     public DateTime ValidFrom { get; set; }
     public DateTime ValidTo { get; set; }
